Fixed delete booking

+++ movie-backend/customer-service/src/main/java/com/brillio/booking/controller/BookingController.java
@@ -3,6 +3,8 @@ package com.brillio.booking.controller;
 import java.util.List;
 import java.util.Optional;
 
+import com.brillio.auth.repository.MovieRepository;
+import com.brillio.booking.model.Movie;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -36,6 +38,9 @@ public class BookingController {
 	@Autowired
 	CustomerRepository customerRepository;
 
+	@Autowired
+	MovieRepository movieRepository;
+
 	@GetMapping("/bookings/{username}")
 	public ResponseEntity<?> getCustomerBookings(@PathVariable String username) {
 		List<CustomerBooking> customerBookings = customerBookingRepository.findAllByUsername(username);
@@ -45,8 +50,24 @@ public class BookingController {
 
 	@DeleteMapping("booking/delete/{id}")
 	public ResponseEntity<String> deleteBooking(@PathVariable String id){
-		customerBookingRepository.deleteById(id);
-		return new ResponseEntity<>("Deleted succesfully", HttpStatus.OK);
+		Optional<CustomerBooking> booking = customerBookingRepository.findById(id);
+		if(booking.isPresent()){
+			String movieId = booking.get().getMovieId();
+			List<Integer> seatsBooked = booking.get().getSeatsBooked();
+
+			customerBookingRepository.deleteById(id);
+			Optional<Movie> movieOptional = movieRepository.findByMovieId(movieId);
+			if(movieOptional.isPresent()){
+				Movie movie = movieOptional.get();
+				List<Integer> occupiedSeats = movie.getOccupiedSeats();
+				occupiedSeats.removeAll(seatsBooked);
+				movie.setOccupiedSeats(occupiedSeats);
+				movieRepository.save(movie);
+			}
+
+			return new ResponseEntity<>("Deleted successfully", HttpStatus.OK);
+		}
+		return new ResponseEntity<>("Could not delete", HttpStatus.PRECONDITION_FAILED);
 	}
 
 	@GetMapping("/users")
